requirements.txt
gymnasium==0.27.1 # no longer supported gym
numpy==1.24.2
tensorflow-macos==2.9.2
tensorflow-metal==0.5.1
keras==2.9.0
# gymnasium[atari] -> python -m pip install gymnasium\[atari\]
autorom==0.4.2
opencv-python==4.7.0.72
moviepy==1.0.3
matplotlib==3.7.1